data-extraction/pyproject.toml
[virtualenvs]
in-project = true

[tool.poetry]
name = "PV Prospect Data Extraction Pipeline"
version = "0.1.0"
description = ""
authors = ["RJ Ferrier <rjferrier@outlook.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.5"
google-api-python-client = "^2.184.0"
google-auth-httplib2 = "==0.2.0"
google-auth-oauthlib = "^1.2.2"
importlib_resources = "^6.5.2"
celery = "^5.5.3"
flower = "^2.0.1"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
ruff = { version = "^0.13" }
mypy = { version = "^1.18" }

[build-system]
requires = ["poetry-core>=2.2.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
